backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
livekit-api==0.7.0
python-dotenv==1.0.1
supabase==2.10.0
python-multipart==0.0.17